move _handle_exception to handlers.py, add flag to enable trapping
