+++ bokeh/application/handlers/code_runner.py
@@ -28,6 +28,7 @@ from types import ModuleType
 
 # Bokeh imports
 from ...util.serialization import make_globally_unique_id
+from .handler import _handle_exception
 
 #-----------------------------------------------------------------------------
 # Globals and constants
@@ -210,16 +211,6 @@ class CodeRunner(object):
 #-----------------------------------------------------------------------------
 # Private API
 #-----------------------------------------------------------------------------
-
-def _handle_exception(handler, e):
-    """helper to stash an exception on a CodeRunner or Handler"""
-    handler._failed = True
-    handler._error_detail = traceback.format_exc()
-
-    _exc_type, _exc_value, exc_traceback = sys.exc_info()
-    filename, line_number, func, txt = traceback.extract_tb(exc_traceback)[-1]
-
-    handler._error = "%s\nFile \"%s\", line %d, in %s:\n%s" % (str(e), os.path.basename(filename), line_number, func, txt)
 
 #-----------------------------------------------------------------------------
 # Code
