+++ bokeh/server/django/routing.py
@@ -68,7 +68,7 @@ class Routing:
 
     def _normalize(self, obj: ApplicationLike) -> Application:
         if callable(obj):
-            return Application(FunctionHandler(obj))
+            return Application(FunctionHandler(obj, trap_exceptions=True))
         elif isinstance(obj, Path):
             return build_single_handler_application(obj)
         else:
